% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFbias.R
\name{BFbias}
\alias{BFbias}
\title{Computes the Bayes Factor for bias}
\usage{
BFbias(y, sig, sig_prior = 2, Z = 1.96)
}
\arguments{
\item{y}{a vector of length n (n=number of studies) representing the effect (standardized mean difference)}

\item{sig}{a vector of length n (n=number of studies) representing the corresponding sampling standard deviation sqrt(variance)}

\item{sig_prior}{a numeric entry representing the standard deviation's prior; default: 2}

\item{Z}{a numeric entry representing the Z value; default: 1.96}
}
\value{
the value of the Bayes Factor
}
\description{
Computes the Bayes Factor for bias
}
\details{
Computes the Bayes Factor for bias in a Meta-Analysis
}
\examples{
library(metafor)
dat <- get(data(dat.bangertdrowns2004))
BFbias(dat$yi, sqrt(dat$vi))
}
\author{
Aldo Cordova-Palomera
}
\seealso{
\code{logL0} \code{logL} \code{nullmass} \code{lmass}
}

