% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logL.R
\name{logL}
\alias{logL}
\title{Logarithm of L, evaluated at specific mu and eta}
\usage{
logL(mu, eta, y, sig, sig_prior = 2, Z = 1.96)
}
\arguments{
\item{mu}{a numeric entry with the value of mu where the expression will be evaluated}

\item{eta}{a numeric entry with the value of eta where the expression will be evaluated}

\item{y}{a vector of length n (n=number of studies) representing the effect (standardized mean difference)}

\item{sig}{a vector of length n (n=number of studies) representing the corresponding sampling standard deviation sqrt(variance)}

\item{sig_prior}{a numeric entry representing the standard deviation's prior; default: 2}

\item{Z}{a numeric entry representing the Z value; default: 1.96}
}
\value{
the value of L evaluated at mu and eta
}
\description{
Logarithm of L, evaluated at specific mu and eta
}
\details{
This is an auxiliary function to calculate the L distribution
in \code{mass}, which is called to compute the Bayes Factor by \code{BFbias}
}
\examples{
library(metafor)
dat <- get(data(dat.bangertdrowns2004))
logL(0, 1, dat$yi, sqrt(dat$vi))
}
\author{
Aldo Córdova-Palomera
}
\seealso{
\code{mass}, \code{BFbias}
}

