% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logL0.R
\name{logL0}
\alias{logL0}
\title{Logarithm of L0, evaluated at a specific mu}
\usage{
logL0(mu, y, sig, sig_prior = 2)
}
\arguments{
\item{mu}{a numeric entry giving the value where the expression will be evaluated}

\item{y}{a vector of length n (n=number of studies) representing the effect (standardized mean difference)}

\item{sig}{a vector of length n (n=number of studies) representing the corresponding sampling standard deviation sqrt(variance)}

\item{sig_prior}{a numeric entry representing the standard deviation's prior; default: 2}
}
\value{
the value of L0 evaluated at mu
}
\description{
Logarithm of L0, evaluated at a specific mu
}
\details{
This is an auxiliary function to calculate the L0 distribution
in \code{nullmass}, which is called to compute the Bayes Factor by \code{BFbias}
}
\examples{
library(metafor)
dat <- get(data(dat.bangertdrowns2004))
logL0(0, dat$yi, sqrt(dat$vi))
}
\author{
Aldo Córdova-Palomera
}
\seealso{
\code{nullmass}, \code{BFbias}
}

