% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mass.R
\name{mass}
\alias{mass}
\title{Value of the definite (two-dimensional) integral of L, 
evaluated around the distribution's mean (for mu), 
and between 0 and 1 (for eta)}
\usage{
mass(y, sig, sig_prior = 2, Z = 1.96)
}
\arguments{
\item{y}{a vector of length n (n=number of studies) representing the effect (standardized mean difference)}

\item{sig}{a vector of length n (n=number of studies) representing the corresponding sampling standard deviation sqrt(variance)}

\item{sig_prior}{a numeric entry representing the standard deviation's prior; default: 2}

\item{Z}{a numeric entry representing the Z value; default: 1.96}
}
\value{
the value of the definite (two-dimensional) integral of the L distribution, 
evaluated around the distribution's mean (+-10 standard deviations, for mu), 
and between 0 and 1 (for eta)
}
\description{
Value of the definite (two-dimensional) integral of L, 
evaluated around the distribution's mean (for mu), 
and between 0 and 1 (for eta)
}
\details{
This is an auxiliary function to obtain the definite integral of L, 
evaluated around the distribution's mean (for mu) and between 0 and 1 for eta. 
It calls the \code{logL} function, and solves the integral using \code{integrate} 
from the \code{cubature} package. 
The output is later called by \code{BFbias} to compute the Bayes Factor.
}
\examples{
library(metafor)
dat <- get(data(dat.bangertdrowns2004))
mass(dat$yi, sqrt(dat$vi))
}
\author{
Aldo Cordova-Palomera
}
\seealso{
\code{logL}, \code{BFbias}
}

