% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullmass.R
\name{nullmass}
\alias{nullmass}
\title{Value of the definite integral of L0, evaluated around the distribution's mean}
\usage{
nullmass(y, sig, sig_prior = 2)
}
\arguments{
\item{y}{a vector of length n (n=number of studies) representing the effect (standardized mean difference)}

\item{sig}{a vector of length n (n=number of studies) representing the corresponding sampling standard deviation sqrt(variance)}

\item{sig_prior}{a numeric entry representing the standard deviation's prior; default: 2}
}
\value{
the value of the definite integral of the L0 distribution, 
evaluated around its mean (+-10 standard deviations)
}
\description{
Value of the definite integral of L0, evaluated around the distribution's mean
}
\details{
This is an auxiliary function to obtain the definite integral of L0, 
evaluated around the distribution's mean. It calls the \code{logL0} function, 
and performs integration using \code{integrate}. 
The output is later called by \code{BFbias} to compute the Bayes Factor.
}
\examples{
library(metafor)
dat <- get(data(dat.bangertdrowns2004))
nullmass(dat$yi, sqrt(dat$vi))
}
\author{
Aldo Cordova-Palomera
}
\seealso{
\code{logL0}, \code{BFbias}
}

